Added Buggy Semi-Functioning Venue GUI (Eventually, and with Messy Code)
Found a method to create the UI without the need for Fragments and instead used recyclerview with the CardLayout view format with clickable elements and placeholder titles. Currently lacks the ability to display up-to-date results and assets from the website, including images and names and to use the tagging system which hopefully will not take too long to add.