Completely refactored Entire Project and added Incomplete Venue GUI
Apologies for the very long delay, researching and adding the GUI took over 1.5 Weeks and also needed to reassemble the project manually due to a failed github push and merge causing the project to become completely broken with a comitt that was unworkable. Was also delayed two days due to issues with Gradle using the wrong libraries and AndroidX automatically overriding the libraries used for the Card UI and causing import failures.