Added messy unfinished CardView (With issues)
Added CardView to the application.  It uses an ugly, hacked in button at the EventOverviewActivity at line 75 to allow the Cardview to be loaded. Due to issues with understanding the JSON Array I could not add the database details and instead used a copied debug ArrayList instead with lots of sloppy code cloned from connor-dev branch. It also can't display images as for some reason the Image Resource Location returns Null for some reason so it was disabled for the time being.