Added New Online Image Downloader With GIF Support [With a Few Bugs]
Managed to eventually add the new Online Image downloader to allow the GridLayout to support Images downloaded Online using a completely new Library Called Glide to do this.

This even includes support for animated Formats wuch as WebP and GIf, However due to a Bug with Glide At the Moment Gif decoding requires a significant Amount of Memeory and CPU and causes large amounts of lag on the Interface if more than a few Gifs are added.

However as Animations are likely Not going to be used in the final product AFAIK The Bug is likely nto a particularly critical one to fix as of now.

The next planned steps I had is to synchronise the App with the Website database with SQL Statements which should not be too hard to do, and to transform each returned row of the databse into a tile representing each venue event. If there is time permitting, I will also try to enable Search functionality and polish teh Interafce with imrpove alignment and posibly New Animation Support.