Additional Cleanup and Fixed Comments
Also the GridRecyclerView class is depreciated (as I did not have the time to fix it), as it used a borked animation handler that is completely broken and causes the GridLayout to crash if used.

It may be work fixing at some point if there is time to do so later.