Added First Implementaion of a Search function and also increased min API level
Allows for searching via matching UserInput with the contents of the Venue ID for all loaded objects. Seems to work suprisingly well as it loads much faster than expected and can even search during typing without needing to manually press a UI element such as a button to do so.

Also increased min API level to fix issues with layout as some of the contraints did not function correctly and to also allow for enhanced features such as animations and the gesture system if needed.

Bugs:

The Keyboard does not dissappear when clicking on the edit text without manually pressing the back button on the phone. (hwoever this does nto seem to effect serach functionality as it loads throughout/During Typing

Animated Gifs Disappear when a SearchQuery is inputted that matches their ID name as displayed in the recyclerview as the Gifs do not seem to be reloaded properly when the Adaptor is reloaded after/During typing