Fixed Bug with Keyboard not Disappearing/Hiding after finished input
Also shortened/added unique ID Names to better demonstrate the search functionality which also helps for easier Searching

Modified Keyboard Input Type to ActionDone to allow the Done Button to Appear on the Keyboard which also fixes the Keyboard hide issue.

Gifs still do not reload properly in the adaptor on inputting a Search Query which may in fact be an issue with Glide and not the App itself.