Fixed Desync Adaptor Bug with Images and Gifs not being loaded correctly by Glide On Search
Was actually a bug caused in CardViewEventAdaptor line 69 where the Default/Unfiltered mData ArraylIst instead of the Filtered Arraylist was being passed to the ViewHolder Upon OnBindViewHolder being Called by the Adaptor, causing the positions of the ArrayList index to be Mixed up and Desync with the Displayed Objects within the Recyclerview and loading the incorrect images as a result.

So in actually the Bug was not caused by Glide at all, and this should fix both of the bugs with the introduction of the Search Function now.